Add user and card information

+++ src/main/java/com/joklek/cardintitysample/shop/impl/ProductImpl.java
@@ -2,15 +2,16 @@ package com.joklek.cardintitysample.shop.impl;
 
 import com.joklek.cardintitysample.shop.Product;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public class ProductImpl implements Product {
 
     private final UUID id;
     private final String title;
-    private final double price;
+    private final BigDecimal price;
 
-    public ProductImpl(UUID id, String title, double price) {
+    public ProductImpl(UUID id, String title, BigDecimal price) {
         this.id = id;
         this.title = title;
         this.price = price;
@@ -27,7 +28,7 @@ public class ProductImpl implements Product {
     }
 
     @Override
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 }
